API and README based documentation, clippy cleanup
