+++ rust-lcm-codec/README.md
@@ -1,0 +1,4 @@
+# rust-lcm-codec
+
+Runtime library support for LCM de/serialization code generated by
+[rust-lcm-codegen](../rust-lcm-codegen).
