+++ rust-lcm-codec/src/lib.rs
@@ -1,7 +1,16 @@
+//! Runtime support for LCM serialization and deserialization in Rust
 #![no_std]
+#![deny(warnings)]
+#![deny(missing_docs)]
+
+/// The errors that can occur when decoding the LCM type hash / fingerprint
+/// for a message.
 #[derive(Debug)]
 pub enum DecodeFingerprintError<E> {
+    /// The fingerprint value found did not match the expected value for the
+    /// message type of interest.
     InvalidFingerprint(u64),
+    /// The underlying StreamingReader encountered an error.
     ReaderError(E),
 }
 
@@ -11,10 +20,17 @@ impl<E> From<E> for DecodeFingerprintError<E> {
     }
 }
 
+/// The errors that can occur when decoding a value in the body
+/// of an LCM message.
 #[derive(Debug)]
 pub enum DecodeValueError<E> {
+    /// The user attempted to read more or fewer items
+    /// out of an array than the array contained.
     ArrayLengthMismatch(&'static str),
+    /// The value attempted to be decoded was invalid
+    /// in some way.
     InvalidValue(&'static str),
+    /// The underlying StreamingReader encountered an error.
     ReaderError(E),
 }
 
@@ -24,10 +40,17 @@ impl<E> From<E> for DecodeValueError<E> {
     }
 }
 
+/// The errors that can occur when encoding a value in the body
+/// of an LCM message.
 #[derive(Debug, PartialEq, Eq)]
 pub enum EncodeValueError<E> {
+    /// The user attempted to write more or fewer items
+    /// into an array than the array contained.
     ArrayLengthMismatch(&'static str),
+    /// The value attempted to be encoded was invalid
+    /// in some way.
     InvalidValue(&'static str),
+    /// The underlying StreamingWriter encountered an error.
     WriterError(E),
 }
 
@@ -39,23 +62,49 @@ impl<E> From<E> for EncodeValueError<E> {
 
 /// Reader backend trait
 pub trait StreamingReader {
+    /// The kind of Error the implementation produces
     type Error;
+    /// Read bytes from the underlying data source into the provided `buf`.
+    /// Should return an error if insufficient bytes are available to
+    /// fully fill `buf`.
     fn read_bytes(&mut self, buf: &mut [u8]) -> Result<(), Self::Error>;
+
+    /// Expose an aliased view of a subset of the underlying data as
+    /// immutable bytes.
+    ///
+    /// The implementer must ensure that the view of bytes returned
+    /// does not overlap with the region of bytes that it allows itself
+    /// to mutate at any further point.
     fn share_bytes(&mut self, len: usize) -> Result<&[u8], Self::Error>;
 }
 
-/// Reader backend for a byte slice
+/// The BufferReader had a problem. The only problem worth mentioning
+/// is that it did not have enough bytes to complete a requested
+/// operations.
 #[derive(Debug)]
 pub struct BufferReaderError;
 
+/// StreamingReader backend for a byte slice
 pub struct BufferReader<'a> {
     buffer: &'a [u8],
     cursor: usize,
 }
 
 impl<'a> BufferReader<'a> {
+    /// Make a new BufferReader around a byte slice
     pub fn new(buffer: &'a [u8]) -> BufferReader<'a> {
         BufferReader { buffer, cursor: 0 }
+    }
+
+    /// How many bytes have been read thus far
+    pub fn cursor(&self) -> usize {
+        self.cursor
+    }
+}
+
+impl<'a> From<&'a [u8]> for BufferReader<'a> {
+    fn from(buffer: &'a [u8]) -> Self {
+        BufferReader::new(buffer)
     }
 }
 
@@ -95,14 +144,27 @@ impl<'a> StreamingReader for BufferReader<'a> {
     }
 }
 
-// Writer backend trait
+/// Writer backend trait
 pub trait StreamingWriter {
+    /// The kind of errors that the implementation emits during encoding
     type Error;
-    // TODO: return size written?
+    /// Write all of the bytes from the provided buffer into the underlying
+    /// encoding stream.
+    ///
+    /// Ought to produce an error if not all of the bytes could be written.
+    ///
+    /// N.B. for possible enhancement: We could return size written here
+    /// rather than leaving that tracking and manner of exposure
+    /// to the implementing type.
     fn write_bytes(&mut self, bytes: &[u8]) -> Result<(), Self::Error>;
+
+    /// Ensure that all bytes are fully written in a maximally durable fashion.
     fn flush() -> Result<(), Self::Error>;
 }
 
+/// The BufferWriter had a problem. The only problem worth mentioning
+/// is that it did not have enough bytes to complete a requested
+/// operations.
 #[derive(Debug, PartialEq, Eq, Copy, Clone)]
 pub struct BufferWriterError;
 
@@ -113,10 +175,12 @@ pub struct BufferWriter<'a> {
 }
 
 impl<'a> BufferWriter<'a> {
+    /// Create a new BufferWriter
     pub fn new(buffer: &'a mut [u8]) -> BufferWriter<'a> {
         BufferWriter { buffer, cursor: 0 }
     }
 
+    /// How many bytes have been written thus far
     pub fn cursor(&self) -> usize {
         self.cursor
     }
@@ -144,9 +208,13 @@ impl<'a> StreamingWriter for BufferWriter<'a> {
     }
 }
 
-// Value serialization trait
+/// Value serialization helper trait, oriented towards primitives.
 pub trait SerializeValue: Sized {
+    /// Use a StreamingReader to produce an instance of the implementing type
+    /// from an encoded stream of LCM data
     fn read_value<R: StreamingReader>(reader: &mut R) -> Result<Self, DecodeValueError<R::Error>>;
+    /// Use a StreamingWriter to write an instance of the implementing type
+    /// to an encoded stream of LCM data
     fn write_value<W: StreamingWriter>(val: Self, writer: &mut W) -> Result<(), W::Error>;
 }
 
@@ -178,12 +246,14 @@ primitive_serialize_impl!(f32);
 primitive_serialize_impl!(f64);
 primitive_serialize_impl!(u8);
 
+/// Write a string to a StreamingWriter using LCM's convention of encoding strings.
 pub fn write_str_value<W: StreamingWriter>(string: &str, writer: &mut W) -> Result<(), W::Error> {
     writer.write_bytes(&(&(string.len() as i32 + 1)).to_be_bytes())?;
     writer.write_bytes(&string.as_bytes())?;
     writer.write_bytes(&[0])
 }
 
+/// Read a view of a string from a StreamingReader using LCM's convention of encoding strings.
 pub fn read_str_value<R: StreamingReader>(
     reader: &mut R,
 ) -> Result<&str, DecodeValueError<<R as StreamingReader>::Error>> {
