+++ rust-lcm-codegen/README.md
@@ -1,0 +1,5 @@
+# rust-lcm-codegen
+
+Generates Rust de/serialization code from [LCM type specification](https://lcm-proj.github.io/type_specification.html)
+files. It is intended to be usable from a Cargo build.rs script.
+
